+++ Almanac/API/ClassesAPI.cs
@@ -6,13 +6,10 @@ namespace Almanac.API;
 public static class ClassesAPI
 {
     private static readonly MethodInfo? API_AddExperience;
-
     public static void AddEXP(int amount)
     {
         API_AddExperience?.Invoke(null, new object[] { amount });
     }
-
-
     static ClassesAPI()
     {
         if (Type.GetType("AlmanacClasses.API.API, AlmanacClasses") is not { } api)